pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wiggy"
version = "0.1.0"
description = "A highly opinionated Ralph Wiggum loop AI software development CLI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Marius Schäffer" },
]
keywords = ["cli", "ai", "development"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.1",
    "docker>=7.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.10",
    "pytest>=8.0",
    "ruff>=0.4",
    "types-docker>=7.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
wiggy = "wiggy.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/wiggy"]

[tool.hatch.build.targets.wheel.force-include]
"src/wiggy/tasks/default" = "wiggy/tasks/default"

[tool.hatch.build.targets.sdist]
include = [
    "src/wiggy",
    "src/wiggy/tasks/default/**/*.yaml",
    "src/wiggy/tasks/default/**/*.md",
]

[tool.ruff]
src = ["src"]
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
